Enemies implemented
The enemies are implemented and should be working. We just need to put
them in a good spot and set the right paths. Collision with the player
may need to be done. For the player it would be nice if we implemented a
boolean to indicate which one is moving.